front/requirements.txt
openpyxl==3.1.5
pandas==2.3.0
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.41.1